Upgrade cranelift to 0.129, pin ariadne, fix all clippy warnings
Cranelift 0.116 → 0.129 required two API migrations: declare_var now
returns the Variable instead of taking one, and jump/brif block args
use BlockArg::Value wrapper. Pinned ariadne to 0.4.1 since 0.6 has
breaking span API changes we don't need yet. Fixed collapsible-if and
redundant-cast clippy warnings in fir and codegen.

629 tests passing, zero warnings.

+++ codegen/src/jit.rs
@@ -228,7 +228,7 @@ impl CraneliftJIT {
                         let local_id = fir::LocalId(i as u32);
                         if let Some(&var) = state.locals.get(&local_id) {
                             let val = builder.use_var(var);
-                            let slot_offset = ((2 + root_idx) * 8) as i32;
+                            let slot_offset = (2 + root_idx) * 8;
                             builder.ins().stack_store(val, gc_frame_slot, slot_offset);
                         }
                         root_idx += 1;
